#576 move runtime banners to runtime_banner.cc
- All banner, no runtime logic.
  Now has it's own little home.
